requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
bcrypt==3.1.6             # via paramiko
certifi==2018.11.29       # via requests
cffi==1.11.5              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
cryptography==2.5         # via paramiko
dateparser==0.7.0
django-dotenv==1.4.2
fabric3==1.14.post1
idna==2.8                 # via requests
paramiko==2.4.2           # via fabric3
pyasn1==0.4.5             # via paramiko
pycparser==2.19           # via cffi
pynacl==1.3.0             # via paramiko
python-dateutil==2.8.0    # via dateparser
pytz==2018.9              # via dateparser, tzlocal
regex==2019.2.7           # via dateparser
requests==2.21.0          # via slackbot, slacker
six==1.12.0               # via bcrypt, cryptography, fabric3, pynacl, python-dateutil, slackbot, websocket-client
slackbot==0.5.3
slacker==0.12.0           # via slackbot
tzlocal==1.5.1            # via dateparser
urllib3==1.24.1           # via requests
websocket-client==0.44.0  # via slackbot
